tests v1 -> v2
rebase tests
